Fix aligning >255, don't align on pointer read/skip

+++ include/binaryio/binaryreader.hpp
@@ -112,7 +112,7 @@ namespace binaryio
 		}
 
 		void Align();
-		void Align(uint_fast8_t byteAlignment);
+		void Align(uint_fast32_t byteAlignment);
 
 		uint64_t ReadPointer();
 		void SkipPointer();
